+++ lib/plugins/jsx/delay/delay.d.ts
@@ -1,0 +1,11 @@
+import { JSXPluginElement } from '@innet/jsx';
+import { Handler } from 'innet';
+import { State } from 'watch-state';
+import { Context } from '../context';
+export interface DelayProps {
+    show?: number;
+    hide?: number;
+}
+export declare const delayContext: Context<State<boolean>, State<boolean>>;
+export declare function useHidden(): undefined | State<boolean>;
+export declare function delay({ props, children }: JSXPluginElement<DelayProps>, handler: Handler): any;