+++ lib/plugins/jsx/slot/slot.d.ts
@@ -3,5 +3,6 @@ import { Handler } from 'innet';
 export interface SlotProps {
     name?: string;
 }
+export declare function useSlots(): Record<string, any>;
 export declare function getSlots(handler: Handler, from: any): Record<string, any>;
 export declare function slot({ props, children }: JSXPluginElement<SlotProps>, handler: any): any;