+++ lib/utils/dom/constants.d.ts
@@ -1,0 +1,1 @@
+export declare const REMOVE_DELAY: unique symbol;