+++ lib/utils/index.d.ts
@@ -3,3 +3,4 @@ export * from './dom';
 export * from './getParent';
 export * from './setParent';
 export * from './useComment';
+export * from './useParent';