+++ lib/utils/useParent/useParent.d.ts
@@ -1,0 +1,2 @@
+import { TargetElements } from '../../types';
+export declare function useParent<T extends TargetElements>(): T;