+++ src/utils/createContext/createContext.ts
@@ -1,3 +1,0 @@
-export function createContext <D> (defaultValue?: D, name?: string): Context {
-
-}