+++ src/utils/dom/constants.ts
@@ -1,0 +1,1 @@
+export const REMOVE_DELAY = Symbol('remove delay')